+++ CHANGELOG.md
@@ -89,6 +89,8 @@ For instructions on upgrading to newer versions, visit
 * \#1381, \#1371 The identity map now functions properly with inherited
   documents. (Paul Canavese)
 
+* \#1366 Array and hash values now get deep copied for dirty tracking.
+
 * \#1359 Provide ability to not have default scope applied to all named
   scopes via using lambdas.